+++ index.html
@@ -230,6 +230,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
+<script src="https://cdn.jsdelivr.net/lodash/4.10.0/lodash.js"></script>
 <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"></script>
 <script src="./js/main.js" type="module"></script>
 </body>
