+++ js/server.js
@@ -8,6 +8,9 @@ let success = document.querySelector('#success').content.querySelector('.success
 let successBlock = success.cloneNode(true);
 let blockMain = document.querySelector('main');
 
+let options = [];
+
+let adFormReset = document.querySelector('.ad-form__reset');
 
 const getReset = function () {
   adFormSubmit.reset();
@@ -21,7 +24,6 @@ const getReset = function () {
 
 adFormSubmit.addEventListener('submit', (evt) => {
   evt.preventDefault();
-
   const formData = new FormData(evt.target);
 
   fetch(
@@ -63,21 +65,7 @@ adFormSubmit.addEventListener('submit', (evt) => {
     });
 
 });
-let options = [];
-let selection = {
-  type: null,
-  price: null,
-  rooms: null,
-  guests: null,
-  wifi: null,
-  dishwasher: null,
-  parking: null,
-  washer: null,
-  elevator: null,
-  conditioner: null,
-}
-let filteredOptions;
-let housingType = document.querySelector('#housing-type');
+
 fetch('https://22.javascript.pages.academy/keksobooking/data')
   .then((response) => response.json())
   .then((cards) => {
@@ -89,170 +77,11 @@ fetch('https://22.javascript.pages.academy/keksobooking/data')
     showAlert('Не удалось связаться с сервером. Попробуйте ещё раз');
   });
 
-let adFormReset = document.querySelector('.ad-form__reset');
+
 adFormReset.addEventListener('click', (evt) => {
   evt.preventDefault();
   getReset();
 })
 
-
-housingType.addEventListener('change', () => {
-  selection.type = housingType.value;
-  filterAndShow();
-})
-let housingRooms = document.querySelector('#housing-rooms');
-housingRooms.addEventListener('change', () => {
-  selection.rooms = housingRooms.value;
-  filterAndShow();
-})
-let housingGuests = document.querySelector('#housing-guests');
-
-housingGuests.addEventListener('change', () => {
-  selection.guests = housingGuests.value;
-  filterAndShow();
-
-})
-let housingPrice = document.querySelector('#housing-price');
-housingPrice.addEventListener('change', () => {
-  selection.price = housingPrice.value;
-  filterAndShow();
-})
-
-let filterWifi = document.querySelector('#filter-wifi');
-let housingFeatures = document.querySelector('#housing-features');
-let featuresInputs = housingFeatures.querySelectorAll('input');
-
-
-featuresInputs[0].addEventListener('change', ()=>{
-  if(!selection.wifi){
-    selection.wifi = filterWifi.value;
-    filterAndShow();
-  }else{
-    selection.wifi = null;
-    filterAndShow();
-  }
-})
-let filterDishwasher = document.querySelector('#filter-dishwasher');
-featuresInputs[1].addEventListener('change', ()=>{
-  if(!selection.dishwasher){
-    selection.dishwasher = filterDishwasher.value;
-    filterAndShow();
-  }else{
-    selection.dishwasher = null;
-    filterAndShow();
-  }
-})
-
-let filterParking = document.querySelector('#filter-parking');
-featuresInputs[2].addEventListener('change', ()=>{
-  if(!selection.parking){
-    selection.parking = filterParking.value;
-    filterAndShow();
-  }else{
-    selection.parking = null;
-    filterAndShow();
-  }
-})
-
-let filterWasher = document.querySelector('#filter-washer');
-featuresInputs[3].addEventListener('change', ()=>{
-  if(!selection.washer){
-    selection.washer = filterWasher.value;
-    filterAndShow();
-  }else{
-    selection.washer = null;
-    filterAndShow();
-  }
-})
-let filterElevator = document.querySelector('#filter-elevator');
-featuresInputs[4].addEventListener('change', ()=>{
-  if(!selection.elevator){
-    selection.elevator = filterElevator.value;
-    filterAndShow();
-  }else{
-    selection.elevator = null;
-    filterAndShow();
-  }
-})
-
-let filterConditioner = document.querySelector('#filter-conditioner');
-featuresInputs[5].addEventListener('change', ()=>{
-  if(!selection.conditioner){
-    selection.conditioner = filterConditioner.value;
-    filterAndShow();
-  }else{
-    selection.conditioner = null;
-    filterAndShow();
-  }
-})
-
-const filterAndShow = function () {
-  filteredOptions = options.slice();
-  if (selection.type) {
-    if (housingType.value !== 'any') {
-      filteredOptions = filteredOptions.filter((o) => o.offer.type === selection.type);
-    }
-  }
-
-  if (selection.rooms) {
-    if (housingRooms.value !== 'any') {
-      filteredOptions = filteredOptions.filter((o) => o.offer.rooms == selection.rooms);
-    }
-  }
-
-  if (selection.guests) {
-    if (housingGuests.value !== 'any') {
-      filteredOptions = filteredOptions.filter((o) => o.offer.guests == selection.guests)
-    }
-  }
-
-  if(selection.price){
-    if (housingPrice.value !== 'any') {
-      if(housingPrice.value === 'middle'){
-        filteredOptions = filteredOptions.filter((o) => o.offer.price > 10000 && o.offer.price <= 50000 );
-      }else if(housingPrice.value === 'low'){
-        filteredOptions = filteredOptions.filter((o) => o.offer.price < 10000 );
-      }else if(housingPrice.value === 'high'){
-        filteredOptions = filteredOptions.filter((o) => o.offer.price > 50000 );
-      }
-    }
-  }
-  if(selection.wifi){
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.wifi) === true);
-  }else{
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.wifi) === false);
-  }
-
-  if(selection.dishwasher){
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.dishwasher) === true);
-  }else{
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.dishwasher) === false);
-  }
-
-  if(selection.parking){
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.parking) === true);
-  }else{
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.parking) === false);
-  }
-
-  if(selection.washer){
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.washer) === true);
-  }else{
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.washer) === false);
-  }
-
-  if(selection.elevator){
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.elevator) === true);
-  }else{
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.elevator) === false);
-  }
-
-  if(selection.conditioner){
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.conditioner) === true);
-  }else{
-    filteredOptions = filteredOptions.filter((o) => o.offer.features.includes(selection.conditioner) === false);
-  }
-  getAdds(filteredOptions.slice(0,10))
-}
-
+export{options}
